# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Alura Project API
  description: API para gerenciamento de simulados, questões e chats de estudo
  version: 1.0.0
servers:
  - url: /v1
    description: Base API path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            status:
              type: integer
...
    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        ratings:
          type: object
...
security:
  - BearerAuth: []
paths:
...
  /questions/{question_id}/chat/start:
    post:
      summary: Iniciar um chat sobre uma questão
      description: Inicia um chat sobre uma questão específica
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: string
          description: ID da questão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Pergunta inicial do usuário
      responses:
        '200':
          description: Chat iniciado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      chat:
                        type: object
                        properties:
                          id:
                            type: string
                          question:
                            $ref: '#/components/schemas/Question'
                          messages:
                            type: array
                            items:
                              type: object
                              properties:
                                content:
                                  type: string
                                isUser:
                                  type: boolean
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Questão não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /questions/chat/{chat_id}/continue:
    post:
      summary: Continuar um chat sobre uma questão
      description: Continua um chat existente sobre uma questão
      security:
        - BearerAuth: []
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: ID do chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Nova pergunta do usuário
      responses:
        '200':
          description: Resposta do chat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      chat:
                        type: object
                        properties:
                          id:
                            type: string
                          question:
                            $ref: '#/components/schemas/Question'
                          messages:
                            type: array
                            items:
                              type: object
                              properties:
                                content:
                                  type: string
                                isUser:
                                  type: boolean
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /questions/chat/history:
    get:
      summary: Obter histórico de chats
      description: Obtém o histórico de chats do usuário autenticado
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de resultados
      responses:
        '200':
          description: Histórico de chats
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      chats:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            questionId:
                              type: string
                            questionText:
                              type: string
                            updatedAt:
                              type: string
                              format: date-time
                            messageCount:
                              type: integer
                      total:
                        type: integer
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /questions/chat/{chat_id}:
    get:
      summary: Obter um chat específico
      description: Obtém os detalhes de um chat específico
      security:
        - BearerAuth: []
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: ID do chat
      responses:
        '200':
          description: Detalhes do chat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      chat:
                        type: object
                        properties:
                          id:
                            type: string
                          question:
                            $ref: '#/components/schemas/Question'
                          messages:
                            type: array
                            items:
                              type: object
                              properties:
                                content:
                                  type: string
                                isUser:
                                  type: boolean
                          createdAt:
                            type: string
                            format: date-time
                          updatedAt:
                            type: string
                            format: date-time
        '403':
          description: Acesso não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'